Separate utils for better tree shaking
