+++ src/utils/if-not-null.js
@@ -1,0 +1,3 @@
+const ifNotNull = (v, alternative = null) => (v === null ? alternative : v);
+
+export default ifNotNull;
