+++ src/utils/unique-str.js
@@ -1,0 +1,5 @@
+import { unique } from '@flekschas/utils';
+
+const uniqueStr = v => unique(v).join('-');
+
+export default uniqueStr;
